simplify rendering
